add tests and resolve comments

+++ src/main/java/com/yelp/nrtsearch/server/luceneserver/warming/WarmerConfig.java
@@ -21,13 +21,13 @@ public class WarmerConfig {
   private static final String CONFIG_PREFIX = "warmer.";
   private static final int DEFAULT_MAX_WARMING_QUERIES = 0;
   private static final int DEFAULT_WARMING_PARALLELISM = 1;
-  private static final float DEFAULT_WARM_BASIC_QUERY_ONLY_PERC = 0.f;
+  private static final int DEFAULT_WARM_BASIC_QUERY_ONLY_PERC = 0;
   private static final boolean DEFAULT_WARM_ON_STARTUP = false;
 
   private final int maxWarmingQueries;
   private final int warmingParallelism;
   private final boolean warmOnStartup;
-  private final float warmBasicQueryOnlyPerc;
+  private final int warmBasicQueryOnlyPerc;
 
   /**
    * Configuration for warmer.
@@ -42,7 +42,7 @@ public class WarmerConfig {
       int maxWarmingQueries,
       int warmingParallelism,
       boolean warmOnStartup,
-      float warmBasicQueryOnlyPerc) {
+      int warmBasicQueryOnlyPerc) {
     this.maxWarmingQueries = maxWarmingQueries;
     this.warmingParallelism = warmingParallelism;
     this.warmOnStartup = warmOnStartup;
@@ -56,8 +56,8 @@ public class WarmerConfig {
         configReader.getInteger(CONFIG_PREFIX + "warmingParallelism", DEFAULT_WARMING_PARALLELISM);
     boolean warmOnStartup =
         configReader.getBoolean(CONFIG_PREFIX + "warmOnStartup", DEFAULT_WARM_ON_STARTUP);
-    float warmBasicQueryOnlyPerc =
-        configReader.getFloat(
+    int warmBasicQueryOnlyPerc =
+        configReader.getInteger(
             CONFIG_PREFIX + "warmBasicQueryOnlyPerc", DEFAULT_WARM_BASIC_QUERY_ONLY_PERC);
 
     return new WarmerConfig(
@@ -76,7 +76,7 @@ public class WarmerConfig {
     return warmOnStartup;
   }
 
-  public float getWarmBasicQueryOnlyPerc() {
+  public int getWarmBasicQueryOnlyPerc() {
     return warmBasicQueryOnlyPerc;
   }
 }
